fix: peer group configs should override only missing peer configs
If GoBGP is configured through the CLI, the peer config takes
precedence over the peer group config because Viper is able to
know what configs are present for each peer. However, when using
the API to define peers, Viper does not read any files and thus
all the peer configs are overwritten by the peer group one.

With the fix, GoBGP also checks if the field is defined in the peer
config before overwritting it with the peer group equivalent.